Acc - basic measure done, scale done, calibration todo

+++ L3GD20.h
@@ -49,7 +49,7 @@ public:
     bool Get_DataAvailableZ();
 
 private:
-    float gain;
+    double gain;
     int maxX;
     int meanX;
     int minX;
